+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## 1.0.0
 
+* [4](https://github.com/cma-arnold/gocd-docker-exec-plugin/issues/4) - Make image used configurable.
 * [3](https://github.com/cma-arnold/gocd-docker-exec-plugin/issues/3) - Get the plugin working with docker.
 * [2](https://github.com/cma-arnold/gocd-docker-exec-plugin/issues/2) - Create inital plugin build.